fix(attack): move attack fn to its module
